+++ pynamodb/expressions/update.py
@@ -1,23 +1,31 @@
+from typing import Dict
 from typing import List
+from typing import Optional
+from typing import Sequence
 from typing import TYPE_CHECKING
 
-from pynamodb.constants import BINARY_SET, NUMBER, NUMBER_SET, STRING_SET
+from pynamodb.constants import BINARY_SET
+from pynamodb.constants import NUMBER
+from pynamodb.constants import NUMBER_SET
+from pynamodb.constants import STRING_SET
 
 if TYPE_CHECKING:
+    from pynamodb.expressions.operand import _Operand
     from pynamodb.expressions.operand import Path
+    from pynamodb.expressions.operand import Value
 
 
-class Action(object):
-    format_string = ''
+class Action:
+    format_string: str = ''
 
-    def __init__(self, *values: 'Path') -> None:
+    def __init__(self, *values: '_Operand') -> None:
         self.values = values
 
-    def serialize(self, placeholder_names, expression_attribute_values):
+    def serialize(self, placeholder_names: Dict[str, str], expression_attribute_values: Dict[str, str]) -> str:
         values = [value.serialize(placeholder_names, expression_attribute_values) for value in self.values]
         return self.format_string.format(*values)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         values = [str(value) for value in self.values]
         return self.format_string.format(*values)
 
@@ -28,7 +36,7 @@ class SetAction(Action):
     """
     format_string = '{0} = {1}'
 
-    def __init__(self, path: 'Path', value: 'Path') -> None:
+    def __init__(self, path: 'Path', value: '_Operand') -> None:
         super(SetAction, self).__init__(path, value)
 
 
@@ -42,22 +50,14 @@ class RemoveAction(Action):
         super(RemoveAction, self).__init__(path)
 
 
-class ListRemoveAction(Action):
-    """
-    The List REMOVE action deletes an element from a list item based on the index.
-    """
-    def __init__(self, path: 'Path', *indexes: int):
-        self.format_string = ", ".join("{{0}}[{}]".format(index) for index in indexes)
-        super(ListRemoveAction, self).__init__(path)
-
-
 class AddAction(Action):
     """
     The ADD action appends elements to a set or mathematically adds to a number attribute.
     """
     format_string = '{0} {1}'
 
-    def __init__(self, path: 'Path', subset: 'Path') -> None:
+    def __init__(self, path: 'Path', subset: 'Value') -> None:
+        path._type_check(BINARY_SET, NUMBER, NUMBER_SET, STRING_SET)
         subset._type_check(BINARY_SET, NUMBER, NUMBER_SET, STRING_SET)
         super(AddAction, self).__init__(path, subset)
 
@@ -68,19 +68,19 @@ class DeleteAction(Action):
     """
     format_string = '{0} {1}'
 
-    def __init__(self, path: 'Path', subset: 'Path') -> None:
+    def __init__(self, path: 'Path', subset: 'Value') -> None:
+        path._type_check(BINARY_SET, NUMBER_SET, STRING_SET)
         subset._type_check(BINARY_SET, NUMBER_SET, STRING_SET)
         super(DeleteAction, self).__init__(path, subset)
 
 
-class Update(object):
+class Update:
 
     def __init__(self, *actions: Action) -> None:
         self.set_actions: List[SetAction] = []
         self.remove_actions: List[RemoveAction] = []
         self.add_actions: List[AddAction] = []
         self.delete_actions: List[DeleteAction] = []
-        self.list_remove_actions: List[ListRemoveAction] = []
         for action in actions:
             self.add_action(action)
 
@@ -89,8 +89,6 @@ class Update(object):
             self.set_actions.append(action)
         elif isinstance(action, RemoveAction):
             self.remove_actions.append(action)
-        elif isinstance(action, ListRemoveAction):
-            self.list_remove_actions.append(action)
         elif isinstance(action, AddAction):
             self.add_actions.append(action)
         elif isinstance(action, DeleteAction):
@@ -98,23 +96,24 @@ class Update(object):
         else:
             raise ValueError("unsupported action type: '{}'".format(action.__class__.__name__))
 
-    def serialize(self, placeholder_names, expression_attribute_values):
-        expression = None
-        expression = self._add_clause(expression, 'SET', self.set_actions, placeholder_names, expression_attribute_values)
-        expression = self._add_clause(expression, 'REMOVE', self.remove_actions, placeholder_names, expression_attribute_values)
-        expression = self._add_clause(expression, 'REMOVE', self.list_remove_actions, placeholder_names, expression_attribute_values)
-        expression = self._add_clause(expression, 'ADD', self.add_actions, placeholder_names, expression_attribute_values)
-        expression = self._add_clause(expression, 'DELETE', self.delete_actions, placeholder_names, expression_attribute_values)
-        return expression
+    def serialize(self, placeholder_names: Dict[str, str], expression_attribute_values: Dict[str, str]) -> Optional[str]:
+        clauses = [
+            self._get_clause('SET', self.set_actions, placeholder_names, expression_attribute_values),
+            self._get_clause('REMOVE', self.remove_actions, placeholder_names, expression_attribute_values),
+            self._get_clause('ADD', self.add_actions, placeholder_names, expression_attribute_values),
+            self._get_clause('DELETE', self.delete_actions, placeholder_names, expression_attribute_values),
+        ]
+        expression = ' '.join(clause for clause in clauses if clause is not None)
+        return expression or None
 
     @staticmethod
-    def _add_clause(expression, keyword, actions, placeholder_names, expression_attribute_values):
-        clause = Update._get_clause(keyword, actions, placeholder_names, expression_attribute_values)
-        if clause is None:
-            return expression
-        return clause if expression is None else expression + " " + clause
-
-    @staticmethod
-    def _get_clause(keyword, actions, placeholder_names, expression_attribute_values):
-        actions = ", ".join([action.serialize(placeholder_names, expression_attribute_values) for action in actions])
-        return keyword + " " + actions if actions else None
+    def _get_clause(
+            keyword: str,
+            actions: Sequence[Action],
+            placeholder_names: Dict[str, str],
+            expression_attribute_values: Dict[str, str]
+    ) -> Optional[str]:
+        actions_string = ', '.join(
+            action.serialize(placeholder_names, expression_attribute_values) for action in actions
+        )
+        return keyword + ' ' + actions_string if actions_string else None