+++ docs/release_notes.rst
@@ -1,15 +1,46 @@
 Release Notes
 =============
 
-v5.0 (unreleased)
------------------
+v5.0.0b1
+-------------------
 
+:date: 2020-xx-xx
+
+This is major release and contains breaking changes. Please read the notes below carefully.
+
+**Polymorphism**
+
+This release introduces :ref:`polymorphism` support via :py:class:`DiscriminatorAttribute <pynamodb.attributes.DiscriminatorAttribute>`.
+Discriminator values are written to DynamoDB and used during deserialization to instantiate the desired class.
+
+** UTCDateTimeAttribute **
+
+The UTCDateTimeAttribute now strictly requires the date string format '%Y-%m-%dT%H:%M:%S.%f%z' to ensure proper ordering.
+PynamoDB has always written values with this format but previously would accept reading other formats.
+Items written using other formats must be rewritten before upgrading.
+
+Other changes in this release:
+
+* Python 2 is no longer supported. Python 3.6 or greater is now required.
 * ``Model.query`` no longer demotes invalid range key conditions to be filter conditions to avoid surprising behaviors:
   where what's intended to be a cheap and fast condition ends up being expensive and slow. Since filter conditions
   cannot contain range keys, this had limited utility to begin with, and would sometimes cause confusing
   "'filter_condition' cannot contain key attributes" errors.
 * Replace the internal attribute type constants with their "short" DynamoDB version (#827)
 * Typed list attributes can now support any Attribute subclass (#833)
+* Add support for empty values in Binary and String attributes (#830)
+* Remove ``ListAttribute.remove_indexes`` (added in v4.3.2) and document usage of remove for list elements (#838)
+* Add the attribute name to error messages when deserialization fails (#815)
+* Add the table name to error messages for transactional operations (#835)
+* Move ``pynamodb.connection.util.pythonic`` to ``pynamodb.util.snake_to_camel_case`` (#753)
+
+Contributors to this release:
+
+* @jpinner
+* @ikonst
+* @rchilaka-amzn
+* @jonathantan
+
 
 v4.3.3
 ----------
@@ -23,7 +54,7 @@ v4.3.3
       my_list = ListAttribute()
 
     MyModel.query(..., condition=MyModel.my_list[0] == 42)
-    
+
 * Fix a warning about ``collections.abc`` deprecation (#782)